+++ NINA.Plugin.SessionMetaData/NINA.Plugin.SessionMetaData/SessionMetaDataWatcher.cs
@@ -296,6 +296,7 @@ namespace SessionMetaData.NINAPlugin {
             public double FocuserTemp { get; set; }
             public double RotatorPosition { get; set; }
             public string PierSide { get; set; }
+            public double Airmass { get; set; }
 
             public ImageMetaDataRecord() {
             }
@@ -335,6 +336,8 @@ namespace SessionMetaData.NINAPlugin {
                 FocuserTemp = Utility.Utility.ReformatDouble(msg.MetaData.Focuser.Temperature);
                 RotatorPosition = Utility.Utility.ReformatDouble(msg.MetaData.Rotator.Position);
                 PierSide = GetPierSide(msg.MetaData.Telescope.SideOfPier);
+
+                Airmass = Utility.Utility.ReformatDouble(msg.MetaData.Telescope.Airmass);
             }
 
             private double GetGuidingMetric(ImageParameter image, double? metric) {