+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Session Metadata
 
 ## 1.3.0.0 - 2002-08-XX
+* Added much better support for specifying the location of metadata files
 * Added output of weather metrics (separate file)
 * Added output of airmass
 * Updated to latest NINA stable dependencies